ipa: simple: awb: Port to use libipa AwbAlgorithm
Port the SoftISP Awb algorithm to use the new libipa implementation
of AwbAlgorithm.

The awbAlgo_ class member is initialized with the Q<2, 8> type even if
there is no physical register representation for SoftISP.

+++ include/libcamera/internal/software_isp/debayer_params.h
@@ -24,7 +24,7 @@ struct DebayerParams {
 	RGB<float> blackLevel = RGB<float>({ 0.0, 0.0, 0.0 });
 	float gamma = 1.0;
 	float contrastExp = 1.0;
-	RGB<float> gains = RGB<float>({ 1.0, 1.0, 1.0 });
+	RGB<double> gains = RGB<double>({ 1.0, 1.0, 1.0 });
 };
 
 } /* namespace libcamera */
